requirements.txt
blessings==1.6.1
certifi==2018.1.18
chardet==3.0.4
curtsies==0.3.0
Django==1.11.11
django-filter==1.1.0
djangorestframework==3.7.7
idna==2.6
Markdown==2.6.11
psycopg2==2.7.4
Pygments==2.2.0
pytz==2018.3
requests==2.18.4
six==1.11.0
typing==3.6.4
urllib3==1.22
wcwidth==0.1.7